Add files via upload
This is an automated Railway announcement system. It requires an excel file containing the train details like train number,train name, station names (from, to, via), platform no. at which train will arrive and the time at which the train is scheduled to arrive. This system will make the announcement 5 minutes before the the train is scheduled to arrive. To generate the announcement audio a pre- recorded announcement audio has been used from which desired portions have been snipped.